pyproject.toml
[project]
name = "graphrag-project"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "neo4j>=5.16.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.3.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "streamlit>=1.30.0",
    "python-multipart>=0.0.6",
    "ollama>=0.1.6",
    "pydantic>=2.5.3",
    "langgraph>=0.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]